refactor: Address PR comments from firestore retriever support
PR: https://github.com/firebase/genkit/pull/2525

CHANGELOG:
  - []
